+++ build/build-v3xctrl.sh
@@ -2,8 +2,24 @@
 
 NAME="v3xctrl"
 
-# Use first argument as ROOT_DIR if provided, otherwise fallback to current working dir
-ROOT_DIR="${1:-$(pwd)}"
+ROOT_DIR="$(pwd)"
+SKIP_DEPS=false
+
+while [[ "$#" -gt 0 ]]; do
+  case $1 in
+    --skip-deps|-s)
+      SKIP_DEPS=true
+      ;;
+    -*)
+      echo "Unknown parameter: $1"
+      exit 1
+      ;;
+    *)
+      ROOT_DIR="$1"
+      ;;
+  esac
+  shift
+done
 
 BUILD_DIR="${ROOT_DIR}/build"
 TMP_DIR="${BUILD_DIR}/tmp"
@@ -21,7 +37,9 @@ PYTHON_LIB_PATH="${DEST_DIR}/opt/v3xctrl-venv/lib/python3.11/site-packages/"
 
 # Clean up previous build (only relevant when re-building on dev setup)
 # In workflows we start with a clean environment anyway
-rm -r "${DEST_DIR}"
+if [ "$SKIP_DEPS" = false ]; then
+  rm -r "${DEST_DIR}"
+fi
 rm "${DEB_PATH}"
 
 # Create dir structure
@@ -43,16 +61,18 @@ cp -r "${ROOT_DIR}/src/v3xctrl_punch" "${PYTHON_LIB_PATH}"
 cp -r "${ROOT_DIR}/src/v3xctrl_udp_relay" "${PYTHON_LIB_PATH}"
 cp -r "${ROOT_DIR}/src/v3xctrl_telemetry" "${PYTHON_LIB_PATH}"
 
-# Fetch static files for the web server
-curl -o "${SERVER_LIB_PATH}/jsoneditor.min.js" "https://raw.githubusercontent.com/jdorn/json-editor/master/dist/jsoneditor.min.js"
-curl -o "${SERVER_LIB_PATH}/jquery.min.js" "https://code.jquery.com/jquery-3.6.0.min.js"
-curl -o "${SERVER_LIB_PATH}/bootstrap3.min.css" "https://cdn.jsdelivr.net/npm/bootstrap@3.3.7/dist/css/bootstrap.min.css"
+if [ "$SKIP_DEPS" = false ]; then
+  # Fetch static files for the web server
+  curl -o "${SERVER_LIB_PATH}/jsoneditor.min.js" "https://raw.githubusercontent.com/jdorn/json-editor/master/dist/jsoneditor.min.js"
+  curl -o "${SERVER_LIB_PATH}/jquery.min.js" "https://code.jquery.com/jquery-3.6.0.min.js"
+  curl -o "${SERVER_LIB_PATH}/bootstrap3.min.css" "https://cdn.jsdelivr.net/npm/bootstrap@3.3.7/dist/css/bootstrap.min.css"
 
-# Install python dependencies
-v3xctrl-pip install \
-  --no-cache-dir \
-  --target "${PYTHON_LIB_PATH}" \
-  -r "${PYTHON_REQUIREMENTS}"
+  # Install python dependencies
+  v3xctrl-pip install \
+    --no-cache-dir \
+    --target "${PYTHON_LIB_PATH}" \
+    -r "${PYTHON_REQUIREMENTS}"
+fi
 
 # Remove cache dirs
 find "${PYTHON_LIB_PATH}" -name '__pycache__' -type d -exec rm -rf {} +
@@ -62,7 +82,7 @@ find "${PYTHON_LIB_PATH}" -name "*.so" -exec strip --strip-unneeded {} \; 2>/dev
 chmod 440 "${DEST_DIR}/etc/sudoers.d/010_v3xctrl"
 
 # Build the deb package
-gzip -9 -n "${DEST_DIR}/usr/share/doc/${NAME}/changelog"
+gzip -9 -n -f "${DEST_DIR}/usr/share/doc/${NAME}/changelog"
 chown -R root:root "${DEST_DIR}"
 
 dpkg-deb --build "${DEST_DIR}" "${DEB_PATH}"
